App menu retirement and fixed deprecated method calls that are causing compiler warnings

See merge request GNOME/sound-juicer!2

+++ src/sj-play.c
@@ -546,7 +546,7 @@ on_volume_changed (GtkWidget * volb, gdouble value, gpointer data)
 {
   vol = value;
   if (pipeline)
-    g_object_set (G_OBJECT (pipeline), "volume", vol, NULL);
+    g_object_set (G_OBJECT (pipeline), "volume", (double) vol, NULL);
   g_settings_set_double (sj_settings, SJ_SETTINGS_AUDIO_VOLUME, vol);
 }
